Add 2 methods that verify player location choice
Write a method to retrieve from a player whether or not their piece
location they chose to move is correct. Essentially they'll enter
yes or no.

Write another method to take this player confirmation and handle it,
returning true if player enters 'yes' or 'y' and false if player
enters anything else.

Adjust main file code to loop back to beginning of player turn if false.