Ensure basic bishop movement is working
Note, there is one test currently pending that does not pass
when working with a white bishop in a certain scenario. However,
white bishops are moving fine on board when running the application.
Most likely is an issue with the test file itself, but cannot figure
it out at this moment.

Noted to take care of later on.