Handle moving own king in check w/ bishop and rook
Now if white moves it's king into the path of a black bishop or rook
with no piece in-between, player will be prompted to choose another
position. Also, if a piece is blocking white king from being in check,
and that piece moves, putting the king into check, then player will
also be prompted to choose another move.