Fix first move bug
Discovered that after one pawn makes it's first move, all other pawns'
instance variables indicating that they've made a first move were set
to true.

Remove the @first_move_made condition as a necessary to check for a
first move, and instead rely on a pawn's current position to know
whether or not it's made a first move already.