Add first iteration of king check move
Currently have a method that prevents any other piece from moving if
king is checked. This does not cover all the cases of getting out of
check, such as moving a piece in front of your king (in the case of
a queen, bishop, or rook checking the king), or taking the piece that's
checking the king.

Will update very soon to handle all these cases.