Fix king check undo board error
When the king was still in check after a player moves one of their
own pieces, the board would revert back to it's previous state and
prompt an error. However, if any enemy pieces were taken during that
turn, their piece would disappear, even when reverting back to the
"original" board.

That error has now been fixed. After a player makes any move, if their
king is still in check an error will be prompted, and the board will
revert back to its previous state properly.