Fix in check pawn error for "pkk error" method
Previously #PKK_error_in_check would return true if a king moved
to a space that the opposing player's pawn could move forward into,
which is a bug. Pawn's can only put a player in check on the following
move if the player puts it's king into the pawn's diagonal, which
should properly throw and error and prompt player to rechoose a move,
and now does so as expected.